pretraining/pretraining_data_preprocessing/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
argcomplete==3.6.1
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.32.0
azure-identity==1.21.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
backoff==2.2.1
bcrypt==4.3.0
boto3==1.37.23
botocore==1.37.23
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
circuitbreaker==2.1.0
contourpy==1.3.1
coolname==2.2.0
cramjam==2.9.1
cryptography==44.0.2
cycler==0.12.1
datasets==3.5.0
dill==0.3.8
duckdb==1.2.1
exceptiongroup==1.2.2
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gql==3.5.2
graphql-core==3.2.4
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
lightning-utilities==0.14.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mosaicml==0.29.0
mosaicml-cli==0.6.42
mosaicml-streaming==0.11.0
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oci==2.149.1
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-snappy==0.7.3
pytorch-ranger==0.1.1
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.4
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
tokenizers==0.21.1
torch==2.6.0
torch-optimizer==0.3.0
torchmetrics==1.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
validators==0.34.0
wcwidth==0.2.13
websockets==11.0.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstd==1.5.6.6
